Refactor distribution inference into master class
