+++ amulet/distribution_inference/attacks/__init__.py
@@ -1,3 +1,3 @@
-from .distribution_inference import DistributionInference
+from .suri_satml_2023 import SuriSATML2023
 
-__all__ = ["DistributionInference"]
+__all__ = ["SuriSATML2023"]
